+++ src/composer.rs
@@ -45,6 +45,11 @@ pub use crate::file_export::export_info::{ExportTrack, SectionInfo, Tone};
 /// placed after the function call are affected; all notes that are already
 /// placed remain unchanged.
 /// 
+/// ## get_active_note
+/// 
+/// Retrieves a mutable reference to the most recent note placed on the track.
+/// The return value is wrapped in an Option in case there are no notes placed.
+/// 
 /// 
 /// ## convert_to_export_track
 /// 
@@ -68,6 +73,8 @@ where
 
     fn set_intensity(&mut self, intensity: f32);
     fn set_play_fraction(&mut self, play_fraction: f32);
+
+    fn get_active_note(&mut self) -> Option<&mut Note<T>>;
 
     fn convert_to_export_track(&self, section_info: SectionInfo) -> ExportTrack<U>;
 }