+++ src/composer/time_signature.rs
@@ -10,13 +10,15 @@ use super::Length;
 /// let four_four =
 ///     TimeSignature::new(4, 4)  // Create a time signature of type 4/4
 ///     .set_beat(0, 1.1)         // Strongly emphasize the first beat
-///     .set_beat(2, 1.05);       // Weakly emphasize the third beat
+///     .set_beat(2, 1.05)        // Weakly emphasize the third beat
+///     .set_offbeat(0.8);        // Weaken offbeats
 /// ```
 #[derive(Clone, PartialEq, Debug)]
 pub struct TimeSignature {
     pub measure_length: Length,
     beat_length: Length,
     beat_intensities: Vec<f32>,
+    offbeat_intensity: f32,
 }
 
 impl TimeSignature {
@@ -38,11 +40,13 @@ impl TimeSignature {
         let beat_length = Length::from_subdivisions(subdivision.into());
         let measure_length = beat_length * nominator.into();
         let beat_intensities = vec![1.0; nominator.into()];
+        let offbeat_intensity = 1.0;
 
         Self {
             measure_length,
             beat_length,
             beat_intensities,
+            offbeat_intensity,
         }
     }
 
@@ -55,6 +59,14 @@ impl TimeSignature {
     /// to the second, etc...
     pub fn set_beat(mut self, index: usize, emphasis: f32) -> Self {
         self.beat_intensities[index] = emphasis;
+        self
+    }
+
+    /// Set the emphasis when no specified beat is hit. This works the same way
+    /// as regular beats, it's just that this emphasis applies when there is no
+    /// beat at the current note.
+    pub fn set_offbeat(mut self, emphasis: f32) -> Self {
+        self.offbeat_intensity = emphasis;
         self
     }
 
@@ -71,6 +83,11 @@ impl TimeSignature {
     /// Return the length of a single beat.
     pub fn beat_length(&self) -> Length {
         self.beat_length
+    }
+
+    /// Return the intensity of an offbeat.
+    pub fn offbeat_intensity(&self) -> f32 {
+        self.offbeat_intensity
     }
 
     fn what_power_of_two(mut value: u8) -> u8 {